+++ src/globals.py
@@ -1,2 +1,2 @@
-build=75
+build=76
 type='dev'